+++ TeamCode/src/main/java/auto/AutoFour.java
@@ -1,14 +1,7 @@
 package auto;
 
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
-import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
-
-import autofunctions.Path;
-import autofunctions.RobotFunctions;
-import autofunctions.TerraCV;
-import global.TerraBot;
-import globalfunctions.Constants;
 
 //@Disabled
 @Autonomous(name="AutoFour", group="Auto")
