+++ TeamCode/src/main/java/auto/AutoHandler.java
@@ -2,17 +2,12 @@ package auto;
 
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 
-import java.io.BufferedWriter;
-import java.io.FileWriter;
-import java.io.IOException;
-
 import autofunctions.Path;
 import autofunctions.RobotFunctions;
 import autofunctions.TerraCV;
 import autofunctions.TerraCVHandler;
 import global.TerraBot;
 import globalfunctions.Constants;
-import globalfunctions.Storage;
 
 public class AutoHandler {
     TerraBot bot = new TerraBot();
@@ -33,52 +28,53 @@ public class AutoHandler {
         path.addWGRF(rf.moveWgTo(55), rf.controlWGE(0.5));
 
         path.addWaypoint(0,10,0);
-        path.addRF(rf.shootRF(3));
-        path.addShoot(0,40, bot);
-
-
+        path.addSetpoint(0,40,0);
+//        path.addRF(rf.shootRF(3));
+//        path.addShoot(0,40, bot);
 //
-//        // Knock down tower, intake 1 ring, and outtake 1 ring
-        path.addRF(rf.intake(1));
-        path.addWaypoint(10,32,0);
-        path.addStop(1.3);
-        path.addRF(rf.shootRF(1));
-        path.addShoot(0,0, bot);
 //
-//        // Intake 3 more rings, outtake them
-        path.addRF(rf.intake(1));
-        for (int i = 0; i < 3; i++) {
-            path.addWaypoint(0, 15, 0);
-            path.addStop(1);
-        }
-        path.addStop(0.5);
-        path.addWaypoint(0,-10,0);
-        path.addRF(rf.shootRF(3));
-        path.addShoot(0,-5, bot);
-        //Place first wobble goal
-        path.addWGRF(rf.controlWGE(1), rf.moveWgTo(0));
-        path.addWaypoint(-55, 115, 0);
-        path.addWGRF(rf.claw(2, -0.2));
-        path.addWaypoint(0,30,0);
-        //Go to 2nd wobble goal
-        path.addWaypoint(10, -20,0);
-        path.addWGRF(rf.moveWgTo(180));
-        path.addWaypoint(70, -120,0);
-        path.addWaypoint(0,-30,0);
-        //Grab it and move to place
-        path.addWGRF(rf.claw(0));
-        path.addStop(1);
-        path.addWGRF(rf.moveWgTo(130));
-        path.addWaypoint(0,20,0);
-        path.addWGRF(rf.controlWGE(1), rf.moveWgTo(0));
-        path.addWaypoint(-80, 115, 0);
-        path.addWGRF(rf.claw(2, -0.2));
-        //Drop 2nd wobble goal
-        path.addWaypoint(0,30,0);
-        path.addStop(0.5);
-        //Park
-        path.addWaypoint(40,-60,0);
-        path.addSetpoint(10,-20,0);
+////
+////        // Knock down tower, intake 1 ring, and outtake 1 ring
+//        path.addRF(rf.intake(1));
+//        path.addWaypoint(10,32,0);
+//        path.addStop(1.3);
+//        path.addRF(rf.shootRF(1));
+//        path.addShoot(0,0, bot);
+////
+////        // Intake 3 more rings, outtake them
+//        path.addRF(rf.intake(1));
+//        for (int i = 0; i < 3; i++) {
+//            path.addWaypoint(0, 15, 0);
+//            path.addStop(1);
+//        }
+//        path.addStop(0.5);
+//        path.addWaypoint(0,-10,0);
+//        path.addRF(rf.shootRF(3));
+//        path.addShoot(0,-5, bot);
+//        //Place first wobble goal
+//        path.addWGRF(rf.controlWGE(1), rf.moveWgTo(0));
+//        path.addWaypoint(-55, 115, 0);
+//        path.addWGRF(rf.claw(2, -0.2));
+//        path.addWaypoint(0,30,0);
+//        //Go to 2nd wobble goal
+//        path.addWaypoint(10, -20,0);
+//        path.addWGRF(rf.moveWgTo(180));
+//        path.addWaypoint(70, -120,0);
+//        path.addWaypoint(0,-30,0);
+//        //Grab it and move to place
+//        path.addWGRF(rf.claw(0));
+//        path.addStop(1);
+//        path.addWGRF(rf.moveWgTo(130));
+//        path.addWaypoint(0,20,0);
+//        path.addWGRF(rf.controlWGE(1), rf.moveWgTo(0));
+//        path.addWaypoint(-80, 115, 0);
+//        path.addWGRF(rf.claw(2, -0.2));
+//        //Drop 2nd wobble goal
+//        path.addWaypoint(0,30,0);
+//        path.addStop(0.5);
+//        //Park
+//        path.addWaypoint(40,-60,0);
+//        path.addSetpoint(10,-20,0);
 
 
 
@@ -87,6 +83,7 @@ public class AutoHandler {
 //        path.saveData();
 
         bot.stopOdoThread();
+        path.stopRFThread();
 
     }
 
@@ -143,6 +140,7 @@ public class AutoHandler {
 //        path.saveData();
 
         bot.stopOdoThread();
+        path.stopRFThread();
     }
 
     public void auto0(){
@@ -160,6 +158,7 @@ public class AutoHandler {
 //        path.saveData();
 //
         bot.stopOdoThread();
+        path.stopRFThread();
     }
 
     public void autoAll(){
