+++ TeamCode/src/main/java/auto/AutoTest.java
@@ -4,12 +4,6 @@ import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 
-import autofunctions.Path;
-import autofunctions.RobotFunctions;
-import autofunctions.TerraCV;
-import global.TerraBot;
-import globalfunctions.Constants;
-
 @Disabled
 @Autonomous(name="AutoTest", group="Auto")
 public class AutoTest extends LinearOpMode {
