+++ TeamCode/src/main/java/auto/AutoZero.java
@@ -4,12 +4,6 @@ import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 
-import autofunctions.Path;
-import autofunctions.RobotFunctions;
-import autofunctions.TerraCV;
-import global.TerraBot;
-import globalfunctions.Constants;
-
 @Disabled
 @Autonomous(name="AutoZero", group="Auto")
 public class AutoZero extends LinearOpMode {
