+++ TeamCode/src/main/java/autofunctions/Path.java
@@ -6,14 +6,19 @@ import com.qualcomm.robotcore.util.Range;
 
 import java.util.ArrayList;
 import java.util.Arrays;
+import java.util.Objects;
 
+import autofunctions.RobotFunctionsHandler;
+import developing.MotionPlanner;
+import global.TerraBot;
+import globalfunctions.Optimizer;
+import globalfunctions.PID;
+import globalfunctions.Sleep;
 import globalfunctions.Storage;
 import globalfunctions.TelemetryHandler;
-import global.TerraBot;
 import globalfunctions.TimeData;
 import util.CodeSeg;
 import util.Line;
-import globalfunctions.PID;
 import util.Vector;
 
 public class Path {
@@ -21,6 +26,10 @@ public class Path {
     public PID xControl = new PID();
     public PID yControl = new PID();
     public PID hControl = new PID();
+
+    public MotionPlanner xMP = new MotionPlanner();
+    public MotionPlanner yMP = new MotionPlanner();
+    public MotionPlanner hMP = new MotionPlanner();
 
     public ElapsedTime globalTime = new ElapsedTime();
     public ElapsedTime endTimer = new ElapsedTime();
@@ -42,9 +51,7 @@ public class Path {
 
     public ArrayList<Double> stops = new ArrayList<>();
     public int stopIndex = 0;
-
     public TelemetryHandler telemetryHandler = new TelemetryHandler();
-
 
     public boolean isExecuting = true;
 
@@ -52,110 +59,79 @@ public class Path {
     final public double[] ds = {0.0003,0.0006,0.0005};
     final public double[] is = {0.000,0.000,0.0000};
 
-    public double[] ksS = {0.02,0.01,0.012};
-    public double[] dsS = {0.0003,0.0008,0.0005};
-//    final public double[] isS = {0.0,0.0,0.0};
+    final public double[] fs = {3,2, 100}; // {0.15,0.15,0.27}; // m/s^2, m/s^2, deg/s^2
+    final public double[] ms = {8,9.4,2500}; // m/s^2, m/s^2, deg/s^2
+    final public double[] ls = {0.6,0.6,0.6}; //s, s, s
 
-    public double[] isS = {0.1,0.08,0.1};
+    public double xp = 0.0;
+    public double yp = 0.0;
+    public double hp = 0.0;
 
-    public double xRestPow = 0.00;
-    public double yRestPow = 0.00;
-    public double hRestPow = 0.02;
-
-
-
-    public double XAcc = 1;
-    public double YAcc = 1;
+    //Accs in meters
+    public double XAcc = 0.01;
+    public double YAcc = 0.01;
     public double HAcc = 1;
-
-
-    final public double endWait = 0.1; // 0.2
 
     public Storage storage = new Storage();
     public ArrayList<double[]> track = new ArrayList<>();
     public ArrayList<double[]> targets = new ArrayList<>();
     public ArrayList<Double> trackTimes = new ArrayList<>();
-
     public ArrayList<double[]> speeds = new ArrayList<>();
-
     public ElapsedTime trackTime = new ElapsedTime();
 
     public boolean globalMode = false;
 
+    public double angleToGoal = 0;
 
+    public double lxpos = 0;
+    public double lypos = 0;
+    public double lhpos = 0;
 
+    public double lxvel = 0;
+    public double lyvel = 0;
+    public double lhvel = 0;
 
+    public double xerr = 0;
+    public double yerr = 0;
+    public double herr = 0;
+
+    public boolean xdone = false;
+    public boolean ydone = false;
+    public boolean hdone = false;
+
+    public double ltime = 0;
+
+    public int updateNum = 0;
 
     public Path(double sx, double sy, double sh){
         poses.add(new double[]{sx, sy, sh});
         init();
     }
-    public Path(double[] pos){
-        poses.add(pos);
+    public Path(double[] spos){
+        poses.add(spos);
         init();
     }
     public void init(){
         posetypes.add(Posetype.SETPOINT);
-        setCoeffsForWaypoint();
-        xControl.setAcc(XAcc);
-        yControl.setAcc(YAcc);
-        hControl.setAcc(HAcc);
-//        xControl.setRestPow(xRestPow);
-//        yControl.setRestPow(yRestPow);
-//        hControl.setRestPow(hRestPow);
-//        xControl.setMaxI(0.15);
-//        yControl.setMaxI(0.15);
-//        hControl.setMaxI(0.15);
+        xControl.setCoefficients(ks[0], ds[0], is[0]);
+        yControl.setCoefficients(ks[1], ds[1], is[1]);
+        hControl.setCoefficients(ks[2], ds[2], is[2]);
         xControl.setMaxD(1);
-        yControl.setMaxD(1); // 0.55
+        yControl.setMaxD(1);
         hControl.setMaxD(1);
-//        xControl.setRangeI(5);
-//        yControl.setRangeI(5);
-//        hControl.setRangeI(5);
+        xMP.setFML(fs[0], ms[0], ls[0]);
+        yMP.setFML(fs[1], ms[1], ls[1]);
+        hMP.setFML(fs[2], ms[2], ls[2]);
+        xMP.setAcc(XAcc);
+        yMP.setAcc(YAcc);
+        hMP.setAcc(HAcc);
+        xMP.setAclP(xp);
+        yMP.setAclP(yp);
+        hMP.setAclP(hp);
         globalTime.reset();
         addStop(0.01);
     }
 
-    public void init2() {
-        xControl.setAcc(XAcc);
-        yControl.setAcc(YAcc);
-        hControl.setAcc(HAcc);
-//        xControl.setRestPow(xRestPow);
-//        yControl.setRestPow(yRestPow);
-//        hControl.setRestPow(hRestPow);
-//        xControl.setMaxI(0.15);
-//        yControl.setMaxI(0.15);
-//        hControl.setMaxI(0.15);
-        xControl.setMaxD(1);
-        yControl.setMaxD(1); // 0.55
-        hControl.setMaxD(1);
-//        xControl.setRangeI(5);
-//        yControl.setRangeI(5);
-//        hControl.setRangeI(5);
-    }
-
-    public void setCoeffsForSetpoint(){
-        xControl.setCoefficients(ksS[0], dsS[0], isS[0]);
-        yControl.setCoefficients(ksS[1], dsS[1], isS[1]);
-        hControl.setCoefficients(ksS[2], dsS[2], isS[2]);
-        xControl.scaleAccs(1);
-        yControl.scaleAccs(1);
-        hControl.scaleAccs(1);
-    }
-    public void setCoeffsForWaypoint(){
-        xControl.setCoefficients(ks[0], ds[0], is[0]);
-        yControl.setCoefficients(ks[1], ds[1], is[1]);
-        hControl.setCoefficients(ks[2], ds[2], is[2]);
-    }
-
-    public void setCoeffsForShoot(){
-        xControl.setCoefficients(ksS[0], dsS[0], isS[0]);
-        yControl.setCoefficients(ksS[1], dsS[1], isS[1]);
-        hControl.setCoefficients(ksS[2], dsS[2], isS[2]);
-        xControl.scaleAccs(1);
-        yControl.scaleAccs(1);
-        hControl.scaleAccs(1);
-    }
     public void updateRadius(double dis){ radius = maxRadius*(1-Math.exp(-(1/maxRadius)*(dis))); }
     public void setGlobalMode(boolean val){
         globalMode = val;
@@ -194,22 +170,12 @@ public class Path {
         wobbleGoalHandler.notRF();
     }
 
-    public void disableIs(){
-//        xControl.disableI = true;
-//        yControl.disableI = true;
-//        hControl.disableI = true;
-    }
 
     public void addShoot(double x, double y, TerraBot bot){
-        double[] last = Arrays.copyOf(poses.get(poses.size()-1), 2);
-        last[0] += x;
-        last[1] += y;
-        addNewPose(x, y, bot.autoAimer.getRobotToGoalAngle(last));
-        //LOOK HERE
+        addNewPose(x, y, 0);
         posetypes.add(Posetype.SHOOT);
         rfsHandler.notRF();
         wobbleGoalHandler.notRF();
-        addSetpoint(0,0,-bot.autoAimer.getRobotToGoalAngle(last));
     }
 
     public void addRF(CodeSeg... segs){
@@ -221,16 +187,17 @@ public class Path {
     }
 
     public void startRFThread(LinearOpMode op){
-        rfsHandler.start(op);
-        wobbleGoalHandler.start(op);
+       rfsHandler.start(op);
+       wobbleGoalHandler.start(op);
     }
     public void stopRFThread(){
-        rfsHandler.stop();
-        wobbleGoalHandler.stop();
+       rfsHandler.stop();
+       wobbleGoalHandler.stop();
     }
 
     public void next(){
         resetControls();
+        updateNum = 0;
         globalTime.reset();
         curIndex++;
         rfsHandler.update();
@@ -278,22 +245,68 @@ public class Path {
     }
 
 
-    public void updateControls(double[] currentPos, double[] target, boolean isShoot){
+    public void updateControlsForWaypoint(double[] currentPos, double[] target){
         double xdis = currentPos[0] - target[0];
         double ydis = currentPos[1] - target[1];
-        double herr;
-        if(!isShoot) {
-            herr = currentPos[2] - poses.get(curIndex + 1)[2];
-        }else{
-            herr = currentPos[2] - target[2];
-        }
-
+        double herr = currentPos[2] - poses.get(curIndex + 1)[2];
         Vector disVect = new Vector(xdis,ydis);
         disVect.rotate(-currentPos[2], Vector.angle.DEGREES);
-
         xControl.update(disVect.x);
         yControl.update(disVect.y);
         hControl.update(herr);
+    }
+
+    public void updateControlsForSetPoint(double[] currentPos, double[] target){
+        xerr = (target[0]-currentPos[0])/100;
+        yerr = (target[1]-currentPos[1])/100;
+
+        Vector errVect = new Vector(xerr,yerr);
+        errVect.rotate(-currentPos[2], Vector.angle.DEGREES);
+
+        xerr = errVect.x;
+        yerr = errVect.y;
+
+        herr = target[2] - currentPos[2];
+        double deltaX = currentPos[0]/100 - lxpos;
+        double deltaY = currentPos[1]/100 - lypos;
+        double deltaH = currentPos[2]- lhpos;
+        lxpos = currentPos[0]/100;
+        lypos = currentPos[1]/100;
+        lhpos = currentPos[2];
+
+        double curTime = globalTime.seconds();
+        double deltaT = curTime-ltime;
+        ltime = curTime;
+
+        double xvel = deltaX/deltaT;
+        double yvel = deltaY/deltaT;
+        double hvel = deltaH/deltaT;
+
+        double deltaXV = xvel-lxvel;
+        double deltaYV = yvel-lyvel;
+        double deltaHV = hvel-lhvel;
+
+        lxvel = xvel;
+        lyvel = yvel;
+        lhvel = hvel;
+
+        double xacl = deltaXV/deltaT;
+        double yacl = deltaYV/deltaT;
+        double hacl = deltaHV/deltaT;
+
+
+
+        if(updateNum > 1){
+            xMP.update(xerr, xvel, xacl);
+            yMP.update(yerr, yvel, yacl);
+            hMP.update(herr, hvel, hacl);
+        }else{
+            updateNum++;
+        }
+
+        if(hasReachedSetpoint(xerr, yerr, herr)){
+            next();
+        }
 
     }
 
@@ -302,6 +315,8 @@ public class Path {
         yControl.reset();
         hControl.reset();
     }
+
+
 
     public void scaleControls(double scale){
         xControl.scaleCoeffs(scale);
@@ -312,18 +327,15 @@ public class Path {
     public double[] update(double[] currentPos, TerraBot bot){
         switch (posetypes.get(curIndex+1)){
             case WAYPOINT:
-                setCoeffsForWaypoint();
-                double[] target = getTargetPos(currentPos);
-                targets.add(target);
-                updateControls(currentPos,target, false);
+                double[] targetW = getTargetPos(currentPos);
+                targets.add(targetW);
+                updateControlsForWaypoint(currentPos, targetW);
                 updateRadius(lines.get(curIndex).getDis());
-                return calcPows();
+                return calcPows(true);
             case SETPOINT:
-                setCoeffsForSetpoint();
-                double[] target1 = poses.get(curIndex+1);
-                updateControls(currentPos,target1, false);
-                hasReachedSetpoint();
-                return calcPows();
+                double[] targetS = poses.get(curIndex+1);
+                updateControlsForSetPoint(currentPos, targetS);
+                return calcPows(false);
             case STOP:
                 if(globalTime.seconds() > stops.get(stopIndex)){
                     next();
@@ -331,29 +343,19 @@ public class Path {
                 }
                 return new double[]{0,0,0};
             case SHOOT:
-                setCoeffsForShoot();
-                double[] target2 = poses.get(curIndex+1);
+                double[] targetSh = poses.get(curIndex+1);
+                targetSh = new double[]{targetSh[0], targetSh[1], angleToGoal};
+                updateControlsForSetPoint(currentPos, targetSh);
                 if(!bot.autoAimer.hasPosBeenUpdated()) {
-                    bot.autoAimer.setOuttakePos(Arrays.copyOf(target2, 2));
+                    bot.autoAimer.setOuttakePos(Arrays.copyOf(targetSh, 2));
+                    angleToGoal = bot.autoAimer.getRobotToGoalAngle(targetSh);
                 }
-                updateControls(currentPos, target2, true);
-
                 if(bot.autoAimer.isDone){
                     bot.autoAimer.ready();
                     next();
                 }
                 bot.outtakeWithCalculations(false);
-                if(xControl.isDone() && yControl.isDone() && hControl.isDone()){
-                    if(endTimer.seconds() > endWait) {
-                        bot.autoAimer.reached();
-                        return new double[]{0, 0, 0};
-                    }else{
-                        return calcPows();
-                    }
-                }else {
-                    endTimer.reset();
-                    return calcPows();
-                }
+                return calcPows(false);
             default:
                 return new double[]{0,0,0};
         }
@@ -362,51 +364,55 @@ public class Path {
 
     }
 
-    public void hasReachedSetpoint(){
-        if(xControl.isDone() && yControl.isDone() && hControl.isDone()){
-            if(endTimer.seconds() > endWait) {
-                next();
-            }
-        }else{
-            endTimer.reset();
-        }
+    public boolean hasReachedSetpoint(double xerr, double yerr, double herr){
+        xdone = xMP.isDone(xerr);
+        ydone = yMP.isDone(yerr);
+        hdone = hMP.isDone(herr);
+        return xdone && ydone && hdone;
     }
 
-    public double[] calcPows(){
-        double[] out = new double[3];
-        out[0] = -Range.clip(xControl.getPower(),-1,1);
-        out[1] = -Range.clip(yControl.getPower(),-1,1);
-        out[2] = -Range.clip(hControl.getPower(),-1,1);
-        return out;
-    }
-
-    public double[] normalize(double[] in){
-        double sum = Math.abs(in[0]) + Math.abs(in[1]) + Math.abs(in[2]);
-        if(sum > 1){
-            return new double[]{in[0]/sum, in[1]/sum, in[2]/sum};
+    public double[] calcPows(boolean isWay){
+        if (isWay) {
+            double[] out = new double[3];
+            out[0] = -Range.clip(xControl.getPower(), -1, 1);
+            out[1] = -Range.clip(yControl.getPower(), -1, 1);
+            out[2] = -Range.clip(hControl.getPower(), -1, 1);
+            return out;
         }else{
-            return in;
+            double[] out = new double[3];
+            out[0] = Range.clip(xMP.getPower(), -1, 1);
+            out[1] = Range.clip(yMP.getPower(), -1, 1);
+            out[2] = Range.clip(hMP.getPower(), -1, 1);
+            return out;
         }
     }
 
 
     public void start(TerraBot bot, LinearOpMode op){
-        globalTime.reset();
-        op.telemetry.addData("Starting", "RF Threads");
+        op.telemetry.addData("Resetting", "Odometry");
         op.telemetry.update();
         telemetryHandler.init(op.telemetry, bot);
         bot.odometry.resetAll(poses.get(0));
         bot.angularPosition.resetGyro(poses.get(0)[2]);
-        trackTime.reset();
+        op.telemetry.addData("Starting", "Movement");
+        op.telemetry.update();
+        globalTime.reset();
+
+
         while (op.opModeIsActive() && isExecuting) {
             double[] pows = update(bot.odometry.getAll(), bot);
-            track.add(bot.odometry.getAll());
-            trackTimes.add(trackTime.seconds());
             bot.move(pows[1], pows[0], pows[2]);
-//            telemetryHandler.addAuton(this, 1);
-            op.telemetry.update();
+//
+//            op.telemetry.addData("xerr", xerr);
+//            op.telemetry.addData("yerr", yerr);
+//            op.telemetry.addData("herr", herr);
+//            op.telemetry.addData("xdone", xdone);
+//            op.telemetry.addData("ydone", ydone);
+//            op.telemetry.addData("hdone", hdone);
+//            op.telemetry.update();
+//
+//            Sleep.trySleep(() -> Thread.sleep(10));
         }
-//        op.telemetry.addData("COMPLETED", Arrays.toString(bot.odometry.getPos()));
         op.telemetry.addData("Status:", "Done");
         op.telemetry.update();
         bot.move(0,0,0);
@@ -415,7 +421,7 @@ public class Path {
 
     public void saveData(){
         storage.saveTimeData(new TimeData("Current", track, trackTimes));
-//        storage.saveTimeData(new TimeData("Current2", speeds, trackTimes));
+        storage.saveTimeData(new TimeData("Current2", speeds, trackTimes));
         storage.saveTimeData(new TimeData("Current3", poses, false));
         storage.saveTimeData(new TimeData("Current4", targets, false));
     }
